+++ release_notes.md
@@ -21,3 +21,4 @@
 ### 🐛 Fixed Issues
 
 - Fix `CVE-2025-48734` by transitive dependency update in `connectivity-ztis`.
+- For OData Generic Client: Fix `disableBufferingHttpResponse()` in `ODataRequestResultGeneric`.